cart_2: Fix after comments
