+++ FakeNFT/Scenes /Cart/Cell/ParentViewController.swift
@@ -1,0 +1,22 @@
+//
+//  ParentViewController.swift
+//  FakeNFT
+//
+//  Created by Alexander Salagubov on 17.09.2024.
+//
+
+import Foundation
+import UIKit
+
+extension UIView {
+    var parentViewController: UIViewController? {
+        var parentResponder: UIResponder? = self
+        while parentResponder != nil {
+            parentResponder = parentResponder?.next
+            if let viewController = parentResponder as? UIViewController {
+                return viewController
+            }
+        }
+        return nil
+    }
+}
